move weather.rb from model to lib directory
